+++ superduper/rest/build.py
@@ -235,10 +235,9 @@ def build_rest_app(app: SuperDuperApp):
     @app.add('/db/show_template', method='get')
     def db_show_template(
         identifier: str,
-        type_id: str = 'template',
         db: 'Datalayer' = DatalayerDependency(),
     ):
-        template: Template = db.load(type_id=type_id, identifier=identifier)
+        template: Template = db.load(type_id='template', identifier=identifier)
         return template.form_template
 
     @app.add('/db/edit', method='get')
@@ -248,10 +247,9 @@ def build_rest_app(app: SuperDuperApp):
         db: 'Datalayer' = DatalayerDependency(),
     ):
         component = db.load(type_id, identifier)
-        template = db.load('template', component['build_template'])
-        template: Template = db.load(type_id=type_id, identifier=identifier)
+        template: Template = db.load('template', component.build_template)
         form = template.form_template
-        form['_variables'] = component['build_variables']
+        form['_variables'] = component.build_variables
         return form
 
     @app.add('/db/metadata/show_jobs', method='get')